+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.query.ts
@@ -37,7 +37,7 @@ const lessonDownloadsQuery =
 
     return lessonDownloadsSchema.parse({
       ...page,
-      ...nextLessons,
+      nextLessons,
       isLegacy: false,
     });
   };